docs: fix suspense fallback trigger by RiddleButton (#52138)
